fix db bugs and move file
